release BE

See merge request s12-webmobile2-sub1/S12P11A807!117

+++ backend/src/main/java/com/dawn/backend/domain/note/service/NoteService.java
@@ -10,6 +10,7 @@ import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 
 import com.dawn.backend.domain.blueprint.entity.BlueprintVersion;
+import com.dawn.backend.domain.blueprint.exception.BlueprintNotFoundException;
 import com.dawn.backend.domain.blueprint.repository.BlueprintVersionRepository;
 import com.dawn.backend.domain.note.dto.BookmarkNoteItem;
 import com.dawn.backend.domain.note.dto.NoteDto;
@@ -20,7 +21,6 @@ import com.dawn.backend.domain.note.dto.request.CreateNoteRequestDto;
 import com.dawn.backend.domain.note.dto.request.GetBookmarkNotesRequestDto;
 import com.dawn.backend.domain.note.dto.request.GetNotesByBlueprintRequestDto;
 import com.dawn.backend.domain.note.dto.request.GetNotesByPinRequestDto;
-import com.dawn.backend.domain.note.dto.request.NoteDetailRequestDto;
 import com.dawn.backend.domain.note.dto.request.UpdateNoteRequestDto;
 import com.dawn.backend.domain.note.dto.response.BookmarkImageResponseDto;
 import com.dawn.backend.domain.note.dto.response.BookmarkNoteResponseDto;
@@ -32,28 +32,40 @@ import com.dawn.backend.domain.note.dto.response.GetNotesByPinResponseDto;
 import com.dawn.backend.domain.note.dto.response.NoteDetailResponseDto;
 import com.dawn.backend.domain.note.dto.response.RecentNoteResponseDto;
 import com.dawn.backend.domain.note.dto.response.UpdateNoteResponseDto;
+import com.dawn.backend.domain.note.entity.EditableNote;
 import com.dawn.backend.domain.note.entity.Note;
 import com.dawn.backend.domain.note.entity.NoteImage;
 import com.dawn.backend.domain.note.entity.UserNoteCheck;
+import com.dawn.backend.domain.note.exception.DeletedNoteException;
+import com.dawn.backend.domain.note.exception.ImageNotFoundException;
+import com.dawn.backend.domain.note.exception.NoteByBlueprintVersionNotFound;
+import com.dawn.backend.domain.note.exception.NoteEditTimeExceededException;
+import com.dawn.backend.domain.note.exception.NoteNotFoundException;
+import com.dawn.backend.domain.note.repository.EditableRepository;
 import com.dawn.backend.domain.note.repository.ImageRepository;
 import com.dawn.backend.domain.note.repository.NoteCheckRepository;
 import com.dawn.backend.domain.note.repository.NoteRepository;
-import com.dawn.backend.domain.pin.dto.ImageItem;
 import com.dawn.backend.domain.pin.dto.PinGroupDto;
 import com.dawn.backend.domain.pin.entity.Pin;
-import com.dawn.backend.domain.pin.entity.PinGroup;
 import com.dawn.backend.domain.pin.entity.PinVersion;
+import com.dawn.backend.domain.pin.exception.PinNotFoundException;
+import com.dawn.backend.domain.pin.exception.PinVersionNotFoundException;
 import com.dawn.backend.domain.pin.repository.PinRepository;
 import com.dawn.backend.domain.pin.repository.PinVersionRepository;
+import com.dawn.backend.domain.project.entity.Project;
+import com.dawn.backend.domain.project.exception.ProjectNotFoundException;
+import com.dawn.backend.domain.project.repository.ProjectRepository;
 import com.dawn.backend.domain.user.dto.ProjectUserDto;
 import com.dawn.backend.domain.user.entity.User;
+import com.dawn.backend.domain.user.entity.UserProject;
 import com.dawn.backend.domain.user.repository.UserProjectRepository;
 import com.dawn.backend.domain.user.repository.UserRepository;
+import com.dawn.backend.global.util.uploader.dto.ImagePathDto;
+import com.dawn.backend.global.util.uploader.service.UploadService;
 
 @Service
 @RequiredArgsConstructor
 public class NoteService {
-
 	private final NoteRepository noteRepository;
 	private final UserProjectRepository userProjectRepository;
 	private final BlueprintVersionRepository blueprintVersionRepository;
@@ -62,14 +74,18 @@ public class NoteService {
 	private final NoteCheckRepository noteCheckRepository;
 	private final UserRepository userRepository;
 	private final PinVersionRepository pinVersionRepository;
+	private final ProjectRepository projectRepository;
+	private final UploadService uploadService;
+	private final EditableRepository editableRepository;
 
 	/**
 	 * 일정 시간마다 note.isDeleted = true 인 노트 삭제 로직 필요
 	 */
 	@Transactional
 	public DeleteNoteResponseDto deleteNote(Long noteId) {
-//		User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
-//		validatePermission(user.getUserId(), noteId);
+		if (!editableRepository.existsById(noteId)) {
+			throw new NoteEditTimeExceededException();
+		}
 
 		Note note = getNoteById(noteId);
 		note.deleteNote();
@@ -79,26 +95,22 @@ public class NoteService {
 
 	@Transactional
 	public UpdateNoteResponseDto updateNote(Long noteId, UpdateNoteRequestDto dto) {
+		if (!editableRepository.existsById(noteId)) {
+			throw new NoteEditTimeExceededException();
+		}
+
 		Note note = getNoteById(noteId);
-
-//		User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
-//		validatePermission(user.getUserId(), noteId);
-
-//		validateNoteCreateTime(note);
-//		validateIsNoteDeleted(note.getIsDeleted());
+		validateIsNoteDeleted(note.getIsDeleted());
 
 		note.updateNoteTitle(dto.noteTitle());
 		note.updateNoteContent(dto.noteContent());
 
 		return new UpdateNoteResponseDto(note.getNoteId());
-
 	}
 
 	@Transactional
 	public BookmarkNoteResponseDto updateBookmarkNote(Long noteId, BookmarkNoteRequestDto dto) {
 		Note note = getNoteById(noteId);
-//		User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
-//		validatePermission(user.getUserId(), noteId);
 
 		updateNoteBookmark(note);
 
@@ -108,8 +120,6 @@ public class NoteService {
 	@Transactional
 	public BookmarkImageResponseDto updateBookmarkImage(Long imageId, BookmarkImageRequestDto dto) {
 		NoteImage noteImage = getNoteImageById(imageId);
-//		User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
-//		validatePermission(user.getUserId(), noteId);
 
 		updateImageBookmark(noteImage);
 
@@ -127,7 +137,7 @@ public class NoteService {
 
 	private NoteImage getNoteImageById(Long imageId) {
 		return imageRepository.findById(imageId)
-			.orElseThrow(() -> new RuntimeException("Image not found"));
+			.orElseThrow(ImageNotFoundException::new);
 	}
 
 	private void updateNoteBookmark(Note note) {
@@ -141,43 +151,30 @@ public class NoteService {
 
 	private void validateIsNoteDeleted(Boolean isDeleted) {
 		if (isDeleted) {
-			throw new RuntimeException("삭제된 노트입니다.");
+			throw new DeletedNoteException();
 		}
-	}
-
-	private void validateNoteCreateTime(Note note) {
-		if (note.getCreatedAt().plusMinutes(5).isBefore(LocalDateTime.now())) {
-			throw new RuntimeException("5분 이후에는 노트를 수정할 수 없습니다.");
-		}
-	}
-
-
-	private void validatePermission(Long userId, Long noteId) {
-		Note note = getNoteById(noteId);
-		Long projectId = note.getBlueprintVersion().getBlueprint().getProject().getProjectId();
-
-		userProjectRepository.findByUserIdAndProjectId(userId, projectId)
-			.orElseThrow(() -> new RuntimeException("User does not have permission to delete this note."));
 	}
 
 	private Note getNoteById(Long noteId) {
 		return noteRepository.findById(noteId)
-			.orElseThrow(() -> new RuntimeException("Note not found with id " + noteId));
+			.orElseThrow(NoteNotFoundException::new);
 	}
 
 	@Transactional
-	public CreateNoteResponseDto createNote(Long userId, Long pinId, CreateNoteRequestDto createNoteRequestDto) {
-
-//		User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
-
-		User user = userRepository.findById(userId)
-			.orElseThrow(() -> new IllegalArgumentException("해당 유저가 존재하지 않습니다."));
+	public CreateNoteResponseDto createNote(User user, Long pinId, CreateNoteRequestDto createNoteRequestDto) {
 
 		BlueprintVersion blueprintVersion
 			= blueprintVersionRepository.findById(createNoteRequestDto.blueprintVersionId())
-			.orElseThrow(() -> new IllegalArgumentException("해당 블루프린트 버전이 존재하지 않습니다."));
+			.orElseThrow(BlueprintNotFoundException::new);
 
-		Pin pin = pinRepository.findById(pinId).orElseThrow(() -> new IllegalArgumentException("해당 핀이 존재하지 않습니다."));
+		Pin pin = pinRepository.findById(pinId)
+			.orElseThrow(PinNotFoundException::new);
+
+		List<User> projectUsers =
+			userProjectRepository.findUserByProjectProjectId(createNoteRequestDto.projectId())
+				.stream()
+				.map(UserProject::getUser)
+				.toList();
 
 		Note note = Note.builder()
 			.noteTitle(createNoteRequestDto.noteTitle())
@@ -187,14 +184,20 @@ public class NoteService {
 			.pin(pin)
 			.build();
 
-		noteRepository.save(note);
+		Note savedNote = noteRepository.save(note);
+		editableRepository.save(new EditableNote(
+			savedNote.getNoteId(),
+			savedNote.getCreatedAt()
+		));
 
 		if (createNoteRequestDto.imageUrlList() != null) {
 			for (String imageUrl : createNoteRequestDto.imageUrlList()) {
+				ImagePathDto imagePathDto = uploadService.getImagePath(imageUrl);
+
 				NoteImage noteImage = NoteImage.builder()
 					.imageOrigin(imageUrl)
-					.imagePreview(imageUrl)
-					.note(note)
+					.imagePreview(imagePathDto.previewPath())
+					.note(savedNote)
 					.build();
 				imageRepository.save(noteImage);
 			}
@@ -207,22 +210,20 @@ public class NoteService {
 		creatorNoteCheck.updateNoteCheck(true);
 		noteCheckRepository.save(creatorNoteCheck);
 
-		List<User> projectUsers = userProjectRepository.findUserByProjectProjectId(createNoteRequestDto.projectId());
-
 		projectUsers.stream()
 			.filter(projectUser -> !projectUser.equals(user))
 			.map(projectUser -> UserNoteCheck.builder()
-				.note(note)
+				.note(savedNote)
 				.user(projectUser)
 				.build())
 			.forEach(noteCheckRepository::save);
 
-		return new CreateNoteResponseDto(note.getNoteId());
+		return new CreateNoteResponseDto(savedNote.getNoteId());
 	}
 
 	public GetNotesByPinResponseDto getNotesByPin(Long pinId, GetNotesByPinRequestDto getNotesByPinRequestDto) {
 		if (!pinRepository.existsById(pinId)) {
-			throw new IllegalArgumentException("해당 핀이 존재하지 않습니다.");
+			throw new PinNotFoundException();
 		}
 
 		List<Note> notes = noteRepository.findAllByPinPinId(pinId);
@@ -255,7 +256,7 @@ public class NoteService {
 
 	public GetBookmarkNotesResponseDto getBookmarkNotes(Long pinId, GetBookmarkNotesRequestDto request) {
 		if (!pinRepository.existsById(pinId)) {
-			throw new IllegalArgumentException("해당 핀이 존재하지 않습니다.");
+			throw new PinNotFoundException();
 		}
 
 		List<Note> bookmarkedNotes = noteRepository.findAllByPinPinIdAndBookmark(pinId, true);
@@ -284,7 +285,7 @@ public class NoteService {
 		Long blueprintId, Long blueprintVersion, GetNotesByBlueprintRequestDto request
 	) {
 		if (!blueprintVersionRepository.existsById(blueprintVersion)) {
-			throw new IllegalArgumentException("해당 블루프린트 버전이 존재하지 않습니다.");
+			throw new NoteByBlueprintVersionNotFound();
 		}
 
 		List<Note> notes = noteRepository.findAllByBlueprintVersion_BlueprintVersionId(blueprintVersion);
@@ -316,39 +317,63 @@ public class NoteService {
 	}
 
 	@Transactional
-	public NoteDetailResponseDto findDetailNote(Long noteId, NoteDetailRequestDto requestDto) {
+	public NoteDetailResponseDto findDetailNote(Long noteId, User user) {
 		Note note = getNoteById(noteId);
-//		User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
-//		validatePermission(user.getUserId(), noteId);
 
 		validateNoteIsDeleted(note);
 
-		User user = userRepository.findById(requestDto.userId())
-			.orElseThrow(() -> new IllegalArgumentException("해당 유저가 존재하지 않습니다."));
-
+		// 노트를 읽음 여부 처리
 		updateUserNoteCheck(note, user);
 
-		Long projectId = note.getBlueprintVersion().getBlueprint().getProject().getProjectId();
-		ProjectUserDto noteWriter = userRepository.findUserWithRoleByUserIdAndProjectId(
-			note.getUser().getUserId(), projectId);
+		//해당 노트를 작성한 유저의 정보 가져오기
+		Project project = getProjectByNoteId(note);
+		ProjectUserDto noteWriter = getNoteWriter(note, project.getProjectId());
 
-		List<NoteImage> noteImages = imageRepository.findAllByNoteNoteIdOrderByBookmark(noteId);
+		List<NoteImage> noteImages = getNoteImages(noteId);
 
-		NoteDto noteDto = NoteDto.from(note, noteWriter, noteImages);
+		boolean isEditable =
+			editableRepository.existsById(note.getNoteId());
 
-		PinVersion pinVersion = pinVersionRepository
-			.findByBlueprintVersionAndPinAndIsActive(note.getBlueprintVersion(), note.getPin(), true)
-			.orElseThrow(() -> new RuntimeException("노트와 연결된 PinVersion 정보를 찾을 수 없습니다. noteId = " + noteId));
-		PinGroup pinGroup = pinVersion.getPinGroup();
+		NoteDto noteDto = NoteDto.from(note, noteWriter, noteImages, isEditable);
 
-		PinGroupDto pinGroupDto = PinGroupDto.from(pinGroup);
+		PinGroupDto pinGroupDto = getPinGroupDto(note, noteId);
 
 		return new NoteDetailResponseDto(noteDto, pinGroupDto);
 	}
 
+	/**
+	 * 노트와 연결된 PinVersion, PinGroup 정보를 조회하고 PinGroupDto로 변환한다.
+	 */
+	private PinGroupDto getPinGroupDto(Note note, Long noteId) {
+		PinVersion pinVersion = pinVersionRepository
+			.findByBlueprintVersionAndPinAndIsActive(note.getBlueprintVersion(), note.getPin(), true)
+			.orElseThrow(PinVersionNotFoundException::new);
+
+		return PinGroupDto.from(pinVersion.getPinGroup());
+	}
+
+	/**
+	 * 노트 이미지 목록을 조회한다.
+	 */
+	private List<NoteImage> getNoteImages(Long noteId) {
+		return imageRepository.findAllByNoteNoteIdOrderByBookmark(noteId);
+	}
+
+	private ProjectUserDto getNoteWriter(Note note, Long projectId) {
+		return userRepository.findUserWithRoleByUserIdAndProjectId(
+			note.getUser().getUserId(),
+			projectId
+		);
+	}
+
+	private Project getProjectByNoteId(Note note) {
+		return projectRepository.findByNoteId(note.getNoteId())
+			.orElseThrow(ProjectNotFoundException::new);
+	}
+
 	private void validateNoteIsDeleted(Note note) {
 		if (note.getIsDeleted()) {
-			throw new RuntimeException("삭제된 노트입니다.");
+			throw new DeletedNoteException();
 		}
 	}
 
@@ -367,16 +392,19 @@ public class NoteService {
 	@Transactional
 	public RecentNoteResponseDto getRecentNoteByPin(Long pinId) {
 		Pin pin = pinRepository.findById(pinId)
-			.orElseThrow(() -> new IllegalArgumentException("해당 핀이 존재하지 않습니다."));
+			.orElseThrow(PinNotFoundException::new);
 
 		Note recentNote = noteRepository.findFirstByPinPinIdAndIsDeletedFalseOrderByCreatedAtDesc(pinId)
-			.orElseThrow(() -> new IllegalArgumentException("해당 핀에 노트가 존재하지 않습니다."));
+			.orElseThrow(NoteNotFoundException::new);
 
 		List<NoteImage> noteImages = imageRepository.findAllByNoteNoteIdOrderByBookmark(recentNote.getNoteId());
 
 		ProjectUserDto writerDto = ProjectUserDto.from(recentNote.getUser(), "USER");
 
-		NoteDto noteDto = NoteDto.from(recentNote, writerDto, noteImages);
+		boolean isEditable =
+			editableRepository.existsById(recentNote.getNoteId());
+
+		NoteDto noteDto = NoteDto.from(recentNote, writerDto, noteImages, isEditable);
 
 		return new RecentNoteResponseDto(noteDto);
 	}
